feat: update authentication and profile components
- Updated auth layout and navigation structure
- Removed deprecated aval.tsx component
- Enhanced auth actions and store implementation
- Added new aval creation functionality
- Improved profile management features
- Updated API structure and logging capabilities
- Updated dependabot configuration
