+++ app/(avales-app)/_layout.tsx
@@ -1,3 +1,4 @@
+import { logger } from "@/core/logger";
 import { useAuthStore } from "@/presentation/auth/store/useAuthStore";
 import { Redirect, Stack } from "expo-router";
 import { useEffect } from "react";
@@ -5,6 +6,7 @@ import { ActivityIndicator, View } from "react-native";
 
 const CheckAuthenticationLayout = () => {
   const { status, checkStatus } = useAuthStore();
+  logger.info("Check authentication status", status);
 
   useEffect(() => {
     checkStatus();
@@ -26,6 +28,15 @@ const CheckAuthenticationLayout = () => {
   return (
     <Stack>
       <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+      <Stack.Screen
+        name="nuevo-aval"
+        options={{
+          presentation: "modal",
+          headerShown: false,
+          gestureEnabled: true,
+          animation: "slide_from_bottom",
+        }}
+      />
     </Stack>
   );
 };
