+++ app/(avales-app)/nuevo-aval.tsx
@@ -1,0 +1,5 @@
+import NuevoAvalScreen from "@/presentation/aval/screens/NuevoAvalScreen";
+
+export default function NuevoAvalModal() {
+  return <NuevoAvalScreen />;
+}
