+++ core/api/api.config.ts
@@ -1,0 +1,15 @@
+import { Platform } from "react-native";
+
+const STAGE = process.env.EXPO_PUBLIC_STAGE || "dev";
+
+export const API_URL =
+  STAGE === "prod"
+    ? process.env.EXPO_PUBLIC_API_URL
+    : Platform.OS === "ios"
+    ? process.env.EXPO_PUBLIC_API_URL_IOS
+    : process.env.EXPO_PUBLIC_API_URL_ANDROID;
+
+export const API_CONFIG = {
+  baseURL: API_URL,
+  timeout: 10000,
+};
