+++ core/auth/actions/auth-action.ts
@@ -1,4 +1,5 @@
-import { avalesApi } from "../api/avalesApi";
+import { httpClient } from "@/core/api/http-client";
+import { logger } from "@/core/logger";
 import { User } from "../interface/user";
 
 export interface AuthResponse {
@@ -28,27 +29,23 @@ const returnUserToken = (
 export const authLogin = async (email: string, password: string) => {
   email = email.trim().toLowerCase();
   try {
-    const { data } = await avalesApi.post<AuthResponse>("/auth/login", {
+    const { data } = await httpClient.post<AuthResponse>("/auth/login", {
       email,
       password,
     });
 
-    console.log("Login successful:", data);
-
     return returnUserToken(data);
   } catch (error) {
-    console.error("Error during login:", error);
     throw new Error("Login failed. Please check your credentials.");
   }
 };
 
 export const authCheckStatus = async () => {
   try {
-    const { data } = await avalesApi.get<AuthResponse>("/auth/check-status");
+    const { data } = await httpClient.get<AuthResponse>("/auth/check-status");
+    logger.info("Check status successful", data);
     return returnUserToken(data);
-  } catch (error) {
-    console.log("Error checking authentication status:", error);
-    // console.error("Error checking auth status:", error);
-    // throw new Error("Failed to check authentication status.");
+  } catch (error: any) {
+    return null;
   }
 };
