+++ core/logger/index.ts
@@ -1,0 +1,9 @@
+import { ReactNativeLogger } from "./logger";
+
+export function getLogger() {
+  return new ReactNativeLogger({
+    name: "avales-app",
+  });
+}
+
+export const logger = getLogger();
